fix: adjust filename and folder after rebase
