--- pkg/gateways/grpc/proto/ledger.proto
@@ -1,48 +1,0 @@
-syntax = "proto3";
-
-option go_package = "github.com/stone-co/the-amazing-ledger/pkg/grpc/proto";
-
-package proto;
-
-import "google/protobuf/empty.proto";
-
-service LedgerService {
-  rpc SaveTransaction(SaveTransactionRequest) returns (google.protobuf.Empty);
-  rpc GetAccountBalance(GetAccountInfoRequest) returns (GetAccountInfoResponse);
-}
-
-// SaveTransactionRequest represents a transaction to be saved. A transaction must
-// have at least two entries, with a valid balance. More info here:
-// https://en.wikipedia.org/wiki/Double-entry_bookkeeping
-message SaveTransactionRequest {
-  string id = 1;              // ID (UUID) to link the entries to a transaction.
-  repeated Entry entries = 2; // The list of entries, where len(entries) must be >= 2.
-}
-
-// Entry represents a new entry on the Ledger.
-message Entry  {
-  string id = 1;               // It's the idempotency key, and must be unique (UUID).
-  string account_id = 2;       // Account involved in the operation.
-  uint64 expected_version = 3; // To deal with optimistic lock.
-  Operation operation = 4;     // Operation: debit or credit.
-  int32 amount = 5;            // Amount (in cents).
-}
-
-// Operation has the possible operations to be used in Entry.
-enum Operation {
-  OPERATION_UNSPECIFIED = 0; // Don't use. It's just the default value.
-  OPERATION_DEBIT = 1;       // Debit operation.
-  OPERATION_CREDIT = 2;      // Credit operation.
-};
-
-// Input account id used like a key
-message GetAccountInfoRequest {
-  string account_id = 1; // Represents an account.
-}
-
-// Output used to represnet version and balance from an account.
-message GetAccountInfoResponse {
-  string account_id = 1; // Represents an account.
-  int64 current_version = 2; // To deal with optimistic lock.
-  int32 balance = 3; // Amount (in cents).
-}
